refactor(dashboard): split index_html.rs into 6 segment modules (#1266)
Closes the last over-cap module. index_html.rs (1869 LOC) is now a
6-part directory module, each part_NN.rs <= 350 LOC.

Structure:
  index_html/
    mod.rs (23 LOC, defines index_html_string() concatenator + test alias)
    part_00.rs (350)
    part_01.rs (350)
    part_02.rs (350)
    part_03.rs (350)
    part_04.rs (350)
    part_05.rs (119)

routes.rs now calls super::index_html::index_html_string().
Tests still use INDEX_HTML via a #[cfg(test)] LazyLock<String>.

After this PR, NO source file in src/ exceeds 400 LOC. #1266 done.
