+++ src/operator_commands_dashboard/index_html/part_05.rs
@@ -1,0 +1,119 @@
+pub(crate) const PART_05: &str = r#"      try {
+        const data = await apiFetch('/api/azlin/tmux-sessions');
+        const hosts = Array.isArray(data.hosts) ? data.hosts : [];
+        if(hosts.length === 0){
+          body.innerHTML = '<div style="color:#8b949e;font-size:.85rem">No configured hosts.</div>';
+        } else {
+          body.innerHTML = hosts.map(h => renderTmuxHost(h)).join('');
+        }
+        const ts = document.getElementById('tmux-last-refreshed');
+        if(ts) ts.textContent = data.refreshed_at ? new Date(data.refreshed_at).toLocaleString() : new Date().toLocaleString();
+      } catch(e) {
+        body.innerHTML = '<div style="color:#f85149;font-size:.85rem">Failed to load tmux sessions: '+esc(e.message||e)+'</div>';
+      }
+    }
+    function renderTmuxHost(h){
+      const host = String(h.host || '');
+      const reachable = !!h.reachable;
+      const sessions = Array.isArray(h.sessions) ? h.sessions : [];
+      const errText = h.error ? String(h.error) : '';
+      const headerColor = reachable ? '#3fb950' : '#f85149';
+      const status = reachable ? '● reachable' : '○ unreachable';
+      let inner;
+      if(!reachable){
+        inner = '<div style="color:#8b949e;font-size:.85rem;padding:.5rem">'
+              + (errText ? esc(errText) : 'host unreachable')
+              + '</div>';
+      } else if(sessions.length === 0){
+        inner = '<div style="color:#8b949e;font-size:.85rem;padding:.5rem">No tmux sessions on this host.</div>';
+      } else {
+        const rows = sessions.map(s => {
+          const name = String(s.name || '');
+          const created = fmtUnixTs(s.created);
+          const attached = s.attached ? '✓' : '—';
+          const wins = (s.windows == null) ? '—' : String(s.windows);
+          const tid = 'tmux-open-'+host+'-'+name;
+          return '<tr>'
+               + '<td style="padding:.3rem .5rem;font-family:monospace">'+esc(name)+'</td>'
+               + '<td style="padding:.3rem .5rem;color:#8b949e">'+esc(created)+'</td>'
+               + '<td style="padding:.3rem .5rem;text-align:center">'+attached+'</td>'
+               + '<td style="padding:.3rem .5rem;text-align:right">'+esc(wins)+'</td>'
+               + '<td style="padding:.3rem .5rem;text-align:right">'
+               +   '<button class="btn" data-testid="'+esc(tid)+'" '
+               +     'onclick="openTmuxAttach('+JSON.stringify(host)+','+JSON.stringify(name)+')">Open</button>'
+               + '</td>'
+               + '</tr>';
+        }).join('');
+        inner = '<table data-testid="tmux-table-'+esc(host)+'" '
+              + 'style="width:100%;border-collapse:collapse;font-size:.88rem">'
+              + '<thead><tr style="border-bottom:1px solid var(--border);color:#8b949e;text-align:left">'
+              + '<th style="padding:.3rem .5rem">Session</th>'
+              + '<th style="padding:.3rem .5rem">Created</th>'
+              + '<th style="padding:.3rem .5rem;text-align:center">Attached?</th>'
+              + '<th style="padding:.3rem .5rem;text-align:right">Windows</th>'
+              + '<th style="padding:.3rem .5rem;text-align:right">Action</th>'
+              + '</tr></thead><tbody>'
+              + rows
+              + '</tbody></table>';
+      }
+      // For unreachable hosts, also expose the host-keyed testid on the wrapper so
+      // e2e tests can find error text without a sessions table.
+      const wrapperTid = reachable ? '' : ' data-testid="tmux-table-'+esc(host)+'"';
+      return '<div'+wrapperTid+' style="margin-top:.6rem;border:1px solid var(--border);border-radius:6px;overflow:hidden">'
+           + '<div style="background:#1a1a2e;padding:.4rem .6rem;display:flex;justify-content:space-between;align-items:center">'
+           +   '<strong style="font-family:monospace">'+esc(host)+'</strong>'
+           +   '<span style="color:'+headerColor+';font-size:.85rem">'+status+'</span>'
+           + '</div>'
+           + inner
+           + '</div>';
+    }
+    function openTmuxAttach(host, session){
+      // Validate identifier shape client-side (mirror of server allow-list).
+      const re = /^[A-Za-z0-9_.-]{1,64}$/;
+      if(!re.test(host) || !re.test(session)){
+        setAgentLogStatus('invalid host or session name', '#f85149');
+        return;
+      }
+      initAgentLogTerminal();
+      if(!agentLogTerm) return;
+      // Tear down any existing agent-log WS before reusing the xterm instance.
+      if(agentLogWS){ try { agentLogWS.close(); } catch(_) {} agentLogWS = null; }
+      agentLogTerm.clear();
+      // Surface the attached target in the existing status row.
+      const nameInput = document.getElementById('agent-log-name');
+      if(nameInput) nameInput.value = host + ':' + session;
+      setAgentLogStatus('attaching to '+host+':'+session+'…', '#d29922');
+      const proto = (window.location.protocol === 'https:') ? 'wss:' : 'ws:';
+      const url = proto + '//' + window.location.host
+                + '/ws/tmux_attach/' + encodeURIComponent(host)
+                + '/' + encodeURIComponent(session);
+      let ws;
+      try { ws = new WebSocket(url); ws.binaryType = 'arraybuffer'; }
+      catch(e){ setAgentLogStatus('connect failed: '+(e&&e.message||e), '#f85149'); return; }
+      agentLogWS = ws;
+      ws.onopen = () => setAgentLogStatus('attached: '+host+':'+session, '#3fb950');
+      ws.onmessage = (ev) => {
+        if(!agentLogTerm) return;
+        if(typeof ev.data === 'string'){
+          agentLogTerm.write(ev.data);
+        } else if(ev.data instanceof ArrayBuffer){
+          const bytes = new Uint8Array(ev.data);
+          // Pass raw bytes through xterm so ANSI escapes render correctly.
+          let s = '';
+          for(let i = 0; i < bytes.length; i++) s += String.fromCharCode(bytes[i]);
+          agentLogTerm.write(s);
+        }
+      };
+      ws.onerror = () => setAgentLogStatus('socket error', '#f85149');
+      ws.onclose = () => { setAgentLogStatus('detached', '#8b949e'); if(agentLogWS === ws) agentLogWS = null; };
+    }
+
+    /* --- Init --- */
+    fetchStatus(); fetchIssues(); fetchDistributed(); fetchAgentOverview();
+    setInterval(fetchAgentOverview,30000);
+    setInterval(fetchStatus,30000);
+    setInterval(fetchIssues,120000);
+  </script>
+</body>
+</html>
+"#;
